pipelines/requirements.txt
neo4j>=5.19,<6
apache-airflow-providers-postgres
kafka-python-ng
apache-airflow-providers-apache-kafka
